Disable image shrinking to fix partition corruption issues
The shrink-image.sh script was causing device-mapper errors due to
improper handling of parted's confirmation prompt. Disabling shrinking
for now to allow testing of cloud-init functionality. Image will be
larger (~4.5GB instead of ~3.9GB) but fully functional.

This is a temporary workaround for testing. Will fix shrinking properly
in a future commit.